+++ sdk/containerregistry/arm-containerregistry/samples/v11-beta/javascript/exportPipelinesListSample.js
@@ -16,7 +16,7 @@ require("dotenv").config();
  * This sample demonstrates how to Lists all export pipelines for the specified container registry.
  *
  * @summary Lists all export pipelines for the specified container registry.
- * x-ms-original-file: specification/containerregistry/resource-manager/Microsoft.ContainerRegistry/preview/2023-08-01-preview/examples/ExportPipelineList.json
+ * x-ms-original-file: specification/containerregistry/resource-manager/Microsoft.ContainerRegistry/preview/2023-11-01-preview/examples/ExportPipelineList.json
  */
 async function exportPipelineList() {
   const subscriptionId =