+++ sdk/containerservice/arm-containerservice/src/models/parameters.ts
@@ -22,7 +22,6 @@ import {
   PrivateEndpointConnection as PrivateEndpointConnectionMapper,
   PrivateLinkResource as PrivateLinkResourceMapper,
   Snapshot as SnapshotMapper,
-  ManagedClusterSnapshot as ManagedClusterSnapshotMapper,
   TrustedAccessRoleBinding as TrustedAccessRoleBindingMapper
 } from "../models/mappers";
 
@@ -53,7 +52,7 @@ export const $host: OperationURLParameter = {
 export const apiVersion: OperationQueryParameter = {
   parameterPath: "apiVersion",
   mapper: {
-    defaultValue: "2023-09-02-preview",
+    defaultValue: "2023-09-01",
     isConstant: true,
     serializedName: "api-version",
     type: {
@@ -183,16 +182,6 @@ export const parameters1: OperationParameter = {
   mapper: TagsObjectMapper
 };
 
-export const ignorePodDisruptionBudget: OperationQueryParameter = {
-  parameterPath: ["options", "ignorePodDisruptionBudget"],
-  mapper: {
-    serializedName: "ignore-pod-disruption-budget",
-    type: {
-      name: "Boolean"
-    }
-  }
-};
-
 export const parameters2: OperationParameter = {
   parameterPath: "parameters",
   mapper: ManagedClusterServicePrincipalProfileMapper
@@ -212,21 +201,6 @@ export const commandId: OperationURLParameter = {
   parameterPath: "commandId",
   mapper: {
     serializedName: "commandId",
-    required: true,
-    type: {
-      name: "String"
-    }
-  }
-};
-
-export const version: OperationURLParameter = {
-  parameterPath: "version",
-  mapper: {
-    constraints: {
-      MaxLength: 24,
-      MinLength: 1
-    },
-    serializedName: "version",
     required: true,
     type: {
       name: "String"
@@ -301,20 +275,6 @@ export const parameters5: OperationParameter = {
   mapper: AgentPoolMapper
 };
 
-export const machineName: OperationURLParameter = {
-  parameterPath: "machineName",
-  mapper: {
-    constraints: {
-      Pattern: new RegExp("^[a-zA-Z0-9][-_a-zA-Z0-9]{0,39}$")
-    },
-    serializedName: "machineName",
-    required: true,
-    type: {
-      name: "String"
-    }
-  }
-};
-
 export const privateEndpointConnectionName: OperationURLParameter = {
   parameterPath: "privateEndpointConnectionName",
   mapper: {
@@ -339,11 +299,6 @@ export const parameters7: OperationParameter = {
 export const parameters8: OperationParameter = {
   parameterPath: "parameters",
   mapper: SnapshotMapper
-};
-
-export const parameters9: OperationParameter = {
-  parameterPath: "parameters",
-  mapper: ManagedClusterSnapshotMapper
 };
 
 export const trustedAccessRoleBindingName: OperationURLParameter = {