+++ sdk/core/core-amqp/test/cbs.spec.ts
@@ -2,7 +2,6 @@
 // Licensed under the MIT license.
 
 import { CbsClient, TokenType, defaultCancellableLock } from "../src";
-import { AbortController } from "@azure/abort-controller";
 import { Connection } from "rhea-promise";
 import { assert } from "chai";
 import { createConnectionStub } from "./utils/createConnectionStub";