+++ sdk/databricks/arm-databricks/src/azureDatabricksManagementClient.ts
@@ -60,7 +60,7 @@ export class AzureDatabricksManagementClient extends coreClient.ServiceClient {
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-databricks/3.0.0`;
+    const packageDetails = `azsdk-js-arm-databricks/3.0.1`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`