+++ sdk/hybridcontainerservice/arm-hybridcontainerservice/src/operations/operations.ts
@@ -14,7 +14,7 @@ import * as Mappers from "../models/mappers";
 import * as Parameters from "../models/parameters";
 import { HybridContainerServiceClient } from "../hybridContainerServiceClient";
 import {
-  ResourceProviderOperation,
+  Operation,
   OperationsListNextOptionalParams,
   OperationsListOptionalParams,
   OperationsListResponse,
@@ -37,7 +37,7 @@ export class OperationsImpl implements Operations {
   /** @param options The options parameters. */
   public list(
     options?: OperationsListOptionalParams
-  ): PagedAsyncIterableIterator<ResourceProviderOperation> {
+  ): PagedAsyncIterableIterator<Operation> {
     const iter = this.listPagingAll(options);
     return {
       next() {
@@ -58,7 +58,7 @@ export class OperationsImpl implements Operations {
   private async *listPagingPage(
     options?: OperationsListOptionalParams,
     settings?: PageSettings
-  ): AsyncIterableIterator<ResourceProviderOperation[]> {
+  ): AsyncIterableIterator<Operation[]> {
     let result: OperationsListResponse;
     let continuationToken = settings?.continuationToken;
     if (!continuationToken) {
@@ -79,7 +79,7 @@ export class OperationsImpl implements Operations {
 
   private async *listPagingAll(
     options?: OperationsListOptionalParams
-  ): AsyncIterableIterator<ResourceProviderOperation> {
+  ): AsyncIterableIterator<Operation> {
     for await (const page of this.listPagingPage(options)) {
       yield* page;
     }
@@ -115,7 +115,7 @@ const listOperationSpec: coreClient.OperationSpec = {
   httpMethod: "GET",
   responses: {
     200: {
-      bodyMapper: Mappers.ResourceProviderOperationList
+      bodyMapper: Mappers.OperationListResult
     },
     default: {
       bodyMapper: Mappers.ErrorResponse
@@ -131,7 +131,7 @@ const listNextOperationSpec: coreClient.OperationSpec = {
   httpMethod: "GET",
   responses: {
     200: {
-      bodyMapper: Mappers.ResourceProviderOperationList
+      bodyMapper: Mappers.OperationListResult
     },
     default: {
       bodyMapper: Mappers.ErrorResponse