+++ sdk/monitor/monitor-opentelemetry/src/types.ts
@@ -4,6 +4,17 @@
 export const AZURE_MONITOR_OPENTELEMETRY_VERSION = "1.1.1";
 export const AZURE_MONITOR_STATSBEAT_FEATURES = "AZURE_MONITOR_STATSBEAT_FEATURES";
 
+/**
+ * Default Breeze endpoint.
+ * @internal
+ */
+export const DEFAULT_BREEZE_ENDPOINT = "https://dc.services.visualstudio.com";
+/**
+ * Default Live Metrics endpoint.
+ * @internal
+ */
+export const DEFAULT_LIVEMETRICS_ENDPOINT = "https://rt.services.visualstudio.com";
+
 export enum StatsbeatFeature {
   NONE = 0,
   DISK_RETRY = 1,