+++ sdk/core/core-client-rest/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.1.6 (Unreleased)
+## 1.2.0 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,14 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+- Upgrade dependency `@azure/abort-controller` to `^2.0.0`.
+
+## 1.1.6 (2023-11-30)
+
+### Features Added
+
+- Add an overload for `createRestError` to accept `PathUncheckedResponse` only.
 
 ## 1.1.5 (2023-11-20)